% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\docType{data}
\name{FACETS}
\alias{FACETS}
\alias{facets}
\title{Facets available for querying cellxgene data}
\format{
\code{FACETS} is an object of class \code{character} of length 8.
}
\usage{
FACETS

facets(cellxgene_db = db(), facets = FACETS)
}
\arguments{
\item{cellxgene_db}{an (optional) cellxgene_db object, as returned
by \code{db()}.}

\item{facets}{a character() vector corersponding to one of the
facets in \code{FACETS}.}
}
\value{
\code{facets()} returns a tibble with columns \code{facet}, \code{label},
\code{ontology_term_id}, and \code{n}, the number of times the facet
label is used in the database.
}
\description{
\code{FACETS} is a character vector of common fields used
to subset cellxgene data.

\code{facets()} is used to query the cellxgene database for
current values of one or all facets.
}
\examples{
f <- facets()

## levels of each facet
f |>
    dplyr::count(facet)

## same as facets(, facets = "organism")
f |>
    dplyr::filter(facet == "organism")

}
\keyword{datasets}
