% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmespath.R
\name{jmespath_version}
\alias{jmespath_version}
\alias{jmespath}
\title{Use JMESpath to query JSON files}
\usage{
jmespath_version()

jmespath(data, path)
}
\arguments{
\item{data}{character(1) JSON string.}

\item{path}{character(1) JMESpath query string.}
}
\value{
\preformatted{`jmespath_version()` returns a character(1) major.minor.patch
version string describing the version of the jsoncons library
on which jmespath is implemented.

`jmespath()` returns a character(1) json string representing
the result of the query.
}
}
\description{
\preformatted{`jmespath_version()` reports the version of the C++ jsoncons
library in use.

`jmespath()` executes a query against a json string using the
'jmespath' specification.
}
}
\details{
\code{jmespath()} is implemented in the jsoncons C++ library.
}
\examples{
jmespath_version()

json <- '{
  "locations": [
    {"name": "Seattle", "state": "WA"},
    {"name": "New York", "state": "NY"},
    {"name": "Bellevue", "state": "WA"},
    {"name": "Olympia", "state": "WA"}
  ]
 }'

jmespath(json, "locations[?state == 'WA'].name | sort(@)") |>
    cat("\n")

}
\seealso{
https://danielaparker.github.io/jsoncons/
}
