% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R, R/datasets.R, R/files.R,
%   R/cellxgene.R
\name{collections_view}
\alias{collections_view}
\alias{datasets_view}
\alias{files_view}
\alias{print.cellxgene_view}
\title{Simplified views of cellxgene collections, datasets, and files}
\usage{
collections_view(cellxgene_db = db())

datasets_view(cellxgene_db = db())

files_view(cellxgene_db = db())

\method{print}{cellxgene_view}(x, ..., with_ids = FALSE)
}
\arguments{
\item{cellxgene_db}{an optional 'cellxgene_db' object, as returned
by \code{db()}.}

\item{x}{an object created by \code{collections_view()},
\code{datasets_view()}, or \code{files_view()}.}

\item{...}{additional arguments required by \code{print()}, but ignored.}

\item{with_ids}{logical(1) indicating whether collection, dataset,
and file ids should be displayed.}
}
\value{
Each function returns a tibble with a simplified
description of the corresponding component of the database.
}
\description{
Simplified views of cellxgene collections, datasets, and files
}
\examples{

db <- db()

collections_view(db)

datasets_view(db)

files_view(db)

}
