% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxg.R
\name{cxg}
\alias{cxg}
\title{Shiny application for discovering, viewing, and downloading
cellxgene data}
\usage{
cxg(as = c("tibble", "sce"))
}
\arguments{
\item{as}{character(1) Return value when quiting the shiny
application. \code{"tibble"} returns a tibble describing selected
datasets (including the location on disk of the downloaded
file). \code{"sce"} returns a list of dataset files imported to R as
SingleCellExperiment objects.}
}
\value{
\code{cxg()} returns either a tibble describing datasets
selected in the shiny application, or a list of datasets
imported into R as SingleCellExperiment objects.
}
\description{
Shiny application for discovering, viewing, and downloading
cellxgene data
}
\examples{
\donttest{
cxg()
}

}
