% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db}
\alias{db}
\title{Retrieve updated cellxgene database metadata}
\usage{
db()
}
\description{
Retrieve updated cellxgene database metadata
}
\details{
The database is retrieved from the cellxgene data portal
web site. 'collections' metadata are retrieved on each call;
metadata on each collection is cached locally for re-use.
}
\examples{
db()

}
