% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R, R/datasets.R, R/files.R
\name{collections}
\alias{collections}
\alias{datasets}
\alias{files}
\title{Query cellxgene collections, datasets, and files}
\usage{
collections(cellxgene_db = db())

datasets(cellxgene_db = db)

files(cellxgene_db = db())
}
\arguments{
\item{cellxgene_db}{an optional 'cellxgene_db' object, as returned
by \code{db()}.}
}
\value{
Each function returns a tibble describing the corresponding
component of the database.
}
\description{
Query cellxgene collections, datasets, and files
}
\examples{
db <- db()

collections(db)

collections(db) |>
    dplyr::glimpse()

datasets(db)

## some datasets have more than one organism, so there are
## additional rows
datasets(db) |>
    dplyr::select(ends_with("_id"), "cell_count", "organism") |>
    tidyr::unnest("organism")

## use unnest_wider (a second level of unnesting) to see the
## organisms in each dataset
datasets(db) |>
    dplyr::select(ends_with("_id"), "cell_count", "organism") |>
    tidyr::unnest("organism") |>
    tidyr::unnest_wider("organism", names_sep = ".")

files(db)

}
